upd: renamed commonUtils to utils
