+++ packages/evm/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./libs/commonUtils";
+export * from "./libs/utils";
 export * from "./libs/constants";
 export * from "./libs/nft";
 export * from "./libs/assets";
